requirements.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Для криптовалютных платежей
tronpy==0.6.1

web3==6.11.0
eth-keys==0.5.1
eth-typing==3.5.0